refactor: Export `LazyExprNamespace`
