+++ narwhals/_ibis/expr_list.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from narwhals._compliant import LazyExprNamespace
 from narwhals._compliant.any_namespace import ListNamespace
-from narwhals._compliant.expr import LazyExprNamespace
 
 if TYPE_CHECKING:
     from narwhals._ibis.expr import IbisExpr
