+++ narwhals/_spark_like/expr_dt.py
@@ -2,8 +2,8 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from narwhals._compliant import LazyExprNamespace
 from narwhals._compliant.any_namespace import DateTimeNamespace
-from narwhals._compliant.expr import LazyExprNamespace
 from narwhals._constants import US_PER_SECOND
 from narwhals._duration import parse_interval_string
 from narwhals._spark_like.utils import (
